Update humlib (64-bit pointer math).

+++ src/hum/humlib.cpp
@@ -1,7 +1,7 @@
 //
 // Programmer:    Craig Stuart Sapp <craig@ccrma.stanford.edu>
 // Creation Date: Sat Aug  8 12:24:49 PDT 2015
-// Last Modified: Sat Sep 12 13:13:02 PDT 2020
+// Last Modified: Sat Sep 12 19:49:01 PDT 2020
 // Filename:      /include/humlib.cpp
 // URL:           https://github.com/craigsapp/humlib/blob/master/src/humlib.cpp
 // Syntax:        C++11
@@ -13019,7 +13019,7 @@ ostream& operator<<(ostream& out, HumHash* hash) {
 
 
 
-typedef long TEMP64BITFIX;
+typedef unsigned long long TEMP64BITFIX;
 
 // declare static variables
 vector<_HumInstrument> HumInstrument::data;
